+++ inject/inject-core/src/test/scala/com/twitter/inject/IntegrationTest.scala
@@ -28,4 +28,6 @@ package com.twitter.inject
  * @see [[com.twitter.inject.Test Finatra Test Class]]
  * @see [[com.twitter.inject.IntegrationTestMixin Finatra IntegrationTestMixin]]
  */
-abstract class IntegrationTest extends Test with IntegrationTestMixin
+abstract class IntegrationTest
+  extends Test
+  with IntegrationTestMixin