+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file. Note that `
 
 ### Added
 
+* inject-core: Remove deprecated `c.t.inject.TestMixin#resetMocks`. Properly
+  use `c.t.inject.Mockito` trait in tests. Deprecate resetting of mocks and
+  resettables in `c.t.inject.IntegrationTestMixin`. ``PHAB_ID=D93876``
+
 * finatra-http: Parameterize `@RouteParam`,`@QueryParam`,`@FormParam`, and 
   `@Header` to allow specifying the field name to read from the params or 
   header map. Previously these annotations only looked for values by the